requirements.txt
numpy==1.19.5
pygame==2.0.1
matplotlib==3.3.3
opencv_python==4.2.0.32
torch==1.3.0
easydict==1.9
Pillow==8.2.0
PyYAML==5.4.1
tensorboardX==2.2
